requirements.txt
boto3==1.34.0
python-dotenv==1.0.0
botocore==1.34.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
psycopg2-binary
